requirements.txt
# FastAPI (Command Side)
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# GraphQL (Read Side)
strawberry-graphql==0.213.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# Kafka
kafka-python==2.0.2
aiokafka==0.10.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# CSV handling
pandas